ci: run more unittests (#14861)
